Remove segment
